CoinTossX-1.1.0 (#2)
* Minor changes

Include a code analysis GutHub action; rename data output file headers

* Update workflows

* Update code-analysis.yml

* Update code-analysis.yml

* Update code-analysis.yml

* Update README.md

* Update GitHub workflows

* Update Client.java

* Revert "Update Client.java"

This reverts commit 85f254152c51e44d8781a0f56fa5f8549cc95272.

* Update Client.java

* Add order cancel and order modify code

* Remove code not required

This has no effect on functionality. Simply the removal of useless code

* Include order cancel functionality

CancelOrderPreProcessor was looking for a price from the order where there was none. This caused a null order list which meant the order was not cancelled

* OrderEntry clientOrderId field test 1

Add the "clientOrderId" field to the OrderEntry class to allow clients to cancel and set their own specific IDs for keeping track of orders

* OrderEntry clientOrderId field test 2

* Clean up

* Update HomePage.html

* LOB snapshot test 1

* LOB snapshot test 2

* Fix LOB snapshot viewer

* LOB snapshot test 3

* Update Client.java

* Fix LOB snapshot viewer

* Improve data output test 1

OrdersSubmitted will now include an extra column for clientOrderId. This is required for order cancels in order to see which order was cancelled.

* Improve data output test 2

* Fix data output

* Improve data output test 4

* Improve data output test 5

* Replace cancel's clientOrderId with origClientOrderId * -1

* Improve data output test 1

* Improve data output test 2

* Fix bugs

Extend LOB depth
Fix handling of trades that walk the LOB

* Fix waitForMarketData method in Client

* Fix the processing of crossing limit orders

* Improve trade execution time accuracy test 1

* Improve trade execution time accuracy test 2

* Improve trade execution time accuracy test 3

* Feature update

Add 2 fields to SymbolStatus - the static and dynamic price references
All logged in clients will receive the static and dynamic price reference at the end of each auction

* Feature update fix

* Fix Original Hawkes code

Allow for Dharmesh's code to conform with the update

* Include updates for Julia and Python code